safari-534.53-branch/LayoutTests/fast/forms/input-number-keyoperation-expected.txt

Test for keyboard operations for <input type=number>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Inserting "ab123cd":
PASS input.value is ""
Press the up arrow key:
PASS input.value is "124"
Press the down arrow key:
PASS input.value is "123"
Disable input element:
PASS input.value is "123"
Read-only input element:
PASS input.value is "123"
PASS successfullyParsed is true

TEST COMPLETE
